apps/backend/requirements.txt
# --- Core backend ---
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings

# --- LangChain core ---
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0

# --- Vector DB (lightweight, disk-based) ---
langchain-chroma>=0.1.0
chromadb>=0.4.24

# --- Local embeddings (FAST, NO TORCH, 512MB SAFE) ---
fastembed>=0.3.0

# --- Google Gemini (TEXT ONLY, free tier) ---
google-genai>=0.4.0

# --- PDF processing (ingestion only) ---
pypdf

# --- Utilities ---
tqdm
